# docs/swagger.yaml
basePath: /api/v1
definitions:
  entity.StageOrder:
    properties:
      order:
        type: integer
      stage_id:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  v1.CreatePipelineRequest:
    properties:
      description:
        type: string
      is_default:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
...
  v1.CreateStageRequest:
    properties:
      color:
        type: string
      is_closed_lost:
        type: boolean
      is_closed_won:
        type: boolean
      name:
        type: string
      probability:
        type: integer
    required:
    - name
    type: object
...
  v1.ReorderStagesRequest:
    properties:
      stages:
        items:
          $ref: '#/definitions/entity.StageOrder'
        type: array
    required:
    - stages
    type: object
...
  v1.UpdatePipelineRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      is_default:
        type: boolean
      name:
        type: string
      probability:
        type: integer
    type: object
  v1.UpdatePipelineStageRequest:
    properties:
      color:
        type: string
      is_closed_lost:
        type: boolean
      is_closed_won:
        type: boolean
      name:
        type: string
      probability:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@crm.com
    name: CRM Support
  description: API untuk sistem CRM (Customer Relationship Management)
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: CRM Project API
  version: "1.0"
paths:
...
  /notifications:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List notifikasi yang belum dibaca
      tags:
      - Notifications
  /notifications/{id}/read:
    patch:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tandai notifikasi sudah dibaca
      tags:
      - Notifications
  /notifications/read-all:
    patch:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tandai semua notifikasi sudah dibaca
      tags:
      - Notifications
  /pipelines:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List semua pipeline
      tags:
      - Pipelines
    post:
      consumes:
      - application/json
      parameters:
      - description: Data pipeline
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CreatePipelineRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Buat pipeline baru
      tags:
      - Pipelines
  /pipelines/{id}:
    delete:
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus pipeline
      tags:
      - Pipelines
    get:
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Detail pipeline
      tags:
      - Pipelines
    patch:
      consumes:
      - application/json
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      - description: Data update pipeline
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdatePipelineRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update pipeline
      tags:
      - Pipelines
  /pipelines/{id}/stages:
    post:
      consumes:
      - application/json
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      - description: Data stage
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CreateStageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tambah stage ke pipeline
      tags:
      - Pipelines
  /pipelines/{id}/stages/{stageId}:
    delete:
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      - description: Stage ID
        in: path
        name: stageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus stage
      tags:
      - Pipelines
    patch:
      consumes:
      - application/json
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      - description: Stage ID
        in: path
        name: stageId
        required: true
        type: string
      - description: Data update stage
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.UpdatePipelineStageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update stage
      tags:
      - Pipelines
  /pipelines/{id}/stages/reorder:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Pipeline ID
        in: path
        name: id
        required: true
        type: string
      - description: Urutan stage baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.ReorderStagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Reorder stages
      tags:
      - Pipelines
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"